fix(web): fix fast refresh
